# Improve token parsing and handling in SQL parser