--- a/src/Carbunqlex/Parsing/SqlKeyword.cs
+++ b/src/Carbunqlex/Parsing/SqlKeyword.cs
@@ -94,6 +94,11 @@ public static class SqlKeyword
 
         CommandKeywords = new HashSet<string>
         {
+            "position",
+            "trim",
+            "both",
+            "trailing",
+            "leading",
             //common
             "as",
             //with