- Introduced switch statements in `ReadOnlyMemoryExtensions.cs` for streamlined token processing.
- Enhanced token reading and state management in `SqlTokenizer.cs` for better performance and readability.
- Analysis speed has improved by about 10%.

## before
| Method                         | Mean       | Error     | StdDev    |
|------------------------------- |-----------:|----------:|----------:|
| Carbunqlex_Parse_Short         |   8.410 μs | 0.0409 μs | 0.0363 μs |
| Carbunqlex_ParseOnly_Short     |   7.537 μs | 0.0313 μs | 0.0261 μs |
| Carbunqlex_Parse_Middle        |  24.017 μs | 0.1175 μs | 0.1041 μs |
| Carbunqlex_ParseOnly_Middle    |  22.081 μs | 0.0659 μs | 0.0514 μs |
| Carbunqlex_Parse_Long          |  44.607 μs | 0.2594 μs | 0.2166 μs |
| Carbunqlex_ParseOnly_Long      |  40.964 μs | 0.3165 μs | 0.2806 μs |
| Carbunqlex_Parse_SuperLong     |  94.404 μs | 0.4854 μs | 0.4053 μs |
| Carbunqlex_ParseOnly_SuperLong |  69.528 μs | 0.2361 μs | 0.1972 μs |


## after
| Method                         | Mean       | Error     | StdDev    |
|------------------------------- |-----------:|----------:|----------:|
| Carbunqlex_Parse_Short         |   7.530 μs | 0.0809 μs | 0.0717 μs |
| Carbunqlex_ParseOnly_Short     |   6.770 μs | 0.0197 μs | 0.0154 μs |
| Carbunqlex_Parse_Middle        |  21.180 μs | 0.0782 μs | 0.0610 μs |
| Carbunqlex_ParseOnly_Middle    |  19.101 μs | 0.1167 μs | 0.1034 μs |
| Carbunqlex_Parse_Long          |  40.113 μs | 0.2124 μs | 0.1774 μs |
| Carbunqlex_ParseOnly_Long      |  38.414 μs | 0.1544 μs | 0.1369 μs |
| Carbunqlex_Parse_SuperLong     |  85.758 μs | 0.8017 μs | 0.7107 μs |
| Carbunqlex_ParseOnly_SuperLong |  62.131 μs | 0.7448 μs | 0.6966 μs |